Move the Pcntl Helper Into a Namespace
Because we'll be adding some more stuff to it.
